backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.39",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9", # Necesario para OAuth2PasswordRequestForm
    "faker>=19.3.1", # Necesario para el seeding de datos
    "redis>=5.0.0", # Cliente Redis para persistencia de blacklist
    "slowapi>=0.1.9", # Rate limiting for API endpoints
    "alembic>=1.17.2",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.24.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "factory-boy>=3.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Pure unit tests (no DB, no Redis, no HTTP)",
    "integration: Tests that need a running database",
    "regression: Tests for previously fixed bugs",
    "slow: Tests that take more than 5 seconds",
]
addopts = "-v --strict-markers"
